Fix IndexKernel.cu build
Fixes `signed-unsigned comparision` warnings introduced by
https://github.com/pytorch/pytorch/pull/104054 that changed type of `num_indices` to unsigned

+++ aten/src/ATen/native/cuda/IndexKernel.cu
@@ -61,8 +61,8 @@ static void launch_kernel(const int64_t N, const func_t& f) {
 template <typename func_t>
 void gpu_index_kernel(TensorIteratorBase& iter, const IntArrayRef index_size, const IntArrayRef index_stride, const func_t& f) {
   const auto num_indices = index_size.size();
-  AT_ASSERT(static_cast<size_t>(num_indices) == index_stride.size());
-  AT_ASSERT(num_indices == iter.ntensors() - 2);
+  AT_ASSERT(num_indices == index_stride.size());
+  AT_ASSERT(static_cast<int64_t>(num_indices) == iter.ntensors() - 2);
 
   if (iter.numel() == 0) {
     return;
@@ -95,7 +95,7 @@ void gpu_index_kernel(TensorIteratorBase& iter, const IntArrayRef index_size, co
 
     int64_t offset = 0;
     #pragma unroll
-    for (int i = 0; i < num_indices; i++) {
+    for (unsigned i = 0; i < num_indices; i++) {
       int64_t index = *(int64_t*)(index_ptrs[i] + offsets[2]);
       CUDA_KERNEL_ASSERT2(-sizes[i] <= index && index < sizes[i] && "index out of bounds");
       if (index < 0) {
